requirements.txt
ultralytics==8.1.0
opencv-python>=4.8.0
torch>=2.0.0
onnx>=1.14.0
onnxruntime>=1.16.0
numpy>=1.24.0
PyYAML>=6.0